+++ assets/levels/Lobby.assets.ron
@@ -1,0 +1,4 @@
+({
+    "Crate": File ( path: "blueprints/Crate.glb" ),
+    "Crate": File ( path: "materials/Crate.glb" ),
+})
